--- installer/mac/copy_x64_binary.py
@@ -1,52 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# Copyright (c) 2020 The Brave Authors. All rights reserved.
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this file,
-# You can obtain one at https://mozilla.org/MPL/2.0/.
-
-import argparse
-import os
-import shutil
-import sys
-
-
-def copy_x64(x64_src_path, x64_dest_path):
-    """Copies the x64 binary from the x64 build dir."""
-
-    if not os.path.exists(x64_src_path):
-        raise Exception('Could not find x64 app (%s)' % x64_src_path)
-
-    if os.path.exists(x64_dest_path):
-        shutil.rmtree(x64_dest_path)
-
-    # TODO check why brave_resources.pak differs between x64 and arm64
-    # Signature artifacts (code signing metadata, Widevine .sig files) are not
-    # Mach-O; lipo cannot merge them. The package step re-signs the universal
-    # app.
-    shutil.copytree(x64_src_path,
-                    x64_dest_path,
-                    symlinks=True,
-                    ignore=shutil.ignore_patterns('Sparkle.framework', '*.pak',
-                                                  '_CodeSignature',
-                                                  'CodeResources', '*.sig'))
-    # remove conflicting files
-    os.remove(os.path.join(x64_dest_path, 'Contents', 'Info.plist'))
-
-
-def main(args):
-    parser = argparse.ArgumentParser(
-        description='Copy macos x64 binary to arm64 for universalize')
-    parser.add_argument('x64_src_path',
-                        help='Source path to the x64 app bundle to copy from.')
-    parser.add_argument(
-        'x64_dest_path',
-        help='Destination path for the x64 app under root_out_dir.')
-    parsed = parser.parse_args(args)
-
-    copy_x64(parsed.x64_src_path, parsed.x64_dest_path)
-
-
-if __name__ == '__main__':
-    sys.exit(main(sys.argv[1:]))
